+++ sdk/storage/azure-storage-blob-cryptography/CHANGELOG.md
@@ -3,10 +3,11 @@
 ## 12.27.0-beta.2 (Unreleased)
 
 ### Features Added
+- Added a new `EncryptionVersion.V2_1` that allows  encrypted blobs to be uploaded using a configurable authenticated region length.
 - Added configuration to allow encrypted blobs to be uploaded using a configurable authenticated region length via 
   `BlobClientSideEncryptionOptions`. The region length can be configured to range between 16 bytes to 1GB. The region 
   length can be set via `BlobClientSideEncryptionOptions.setAuthenticatedRegionDataLengthInBytes(long authenticatedRegionDataLength)`. 
-  Note: This change only applies `EncryptionVersion.V2`. Also, only applies to upload operations, this does not directly 
+  Note: This change only applies to `EncryptionVersion.V2_1`. Also, only applies to upload operations, this does not directly 
   change the authenticated region length used to download and decrypt blobs.
 ### Breaking Changes